pyproject.toml
[tool.poetry]
name = "mdnminiblogassessment"
version = "0.1.0"
description = ""
authors = ["GSPVK"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
pyjwt = "^2.8.0"
coverage = "^7.5.3"
dj-database-url = "^2.2.0"
django-ckeditor = "^6.7.1"
django-cors-headers = "^4.3.1"
django-debug-toolbar = "^4.4.2"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-filter = "^24.2"
django-jazzmin = "^3.0.0"
django-phonenumber-field = "^7.3.0"
django-phonenumbers = "^1.0.1"
django-resized = "^1.0.2"
django-schema-graph = "^3.1.0"
django-tinymce = "^4.0.0"
djangoviz = "^0.1.1"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
graphviz = "^0.20.3"
pillow = "^10.3.0"
psycopg2 = "^2.9.9"
pydot = "^2.0.0"
pytz = "^2024.1"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
flower = "^2.0.1"
redis = {extras = ["hiredis"], version = "^5.0.7"}
django-bleach = "^3.1.0"
gunicorn = "^22.0.0"
freezegun = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
